Feat: image name validation when upload file

+++ models/aws_sdk.ts
@@ -1,3 +1,4 @@
+import { verifyingStr } from "@/utils/validation";
 import { S3Client, PutObjectCommand } from "@aws-sdk/client-s3";
 import formidable from "formidable";
 import fs from "fs/promises";
@@ -16,20 +17,27 @@ const s3 = new S3Client({
  *
  *  ex) filename.jpg로 반환
  */
-const uploadFile = async (file: formidable.File) => {
+const uploadFile = async (
+  file: formidable.File,
+  name: string,
+  type: string
+) => {
   try {
+    if (!verifyingStr(name)) throw new Error("Invalid file name");
+    if (!(type === "image/jpeg" || type === "image/png"))
+      throw new Error("Invalid file type");
     const buffer = await fs.readFile(file.filepath);
-
+    let key = name + "." + type.replace(/(image\/)/g, "");
     const uploadParams = {
       Bucket: S3_BUCKET,
-      Key: file.originalFilename || "",
+      Key: key || "",
       Body: buffer,
       ContentType: file.mimetype || "",
     };
 
     await s3.send(new PutObjectCommand(uploadParams));
     await fs.unlink(file.filepath);
-    return uploadParams.Key;
+    return name;
   } catch (err) {
     console.error(err);
     await fs.unlink(file.filepath);
